% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{police_effort_index}
\alias{police_effort_index}
\title{Police effort index}
\format{
A data frame with 10 rows and 3 variables
}
\usage{
police_effort_index
}
\description{
Police effort index
}
\keyword{datasets}
