% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crime_journal_notes}
\alias{crime_journal_notes}
\title{Crime journal notes}
\format{
A character vector with 10 observations
}
\source{
Synthetic data generated for teaching
}
\usage{
crime_journal_notes
}
\description{
A character vector of fictional citizen notes related to crime concerns.
}
\keyword{datasets}
