% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crime_stats_ghent}
\alias{crime_stats_ghent}
\title{Crime statistics in fictional districts}
\format{
A data frame with 60 rows and 5 variables:
\describe{
  \item{district}{Fictional district name}
  \item{year}{Year}
  \item{violent_crimes}{Number of violent crimes}
  \item{property_crimes}{Number of property crimes}
  \item{population}{Total population}
}
}
\usage{
crime_stats_ghent
}
\description{
A dataset with synthetic police-recorded crime data.
}
\keyword{datasets}
