% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fear_of_crime_survey}
\alias{fear_of_crime_survey}
\title{Survey on fear of crime}
\format{
A data frame with 200 rows and 5 variables
}
\usage{
fear_of_crime_survey
}
\description{
Survey on fear of crime
}
\keyword{datasets}
