% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{neighborhood_index}
\alias{neighborhood_index}
\title{Socioeconomic indicators by neighborhood}
\format{
A data frame with 10 rows and 4 variables
}
\usage{
neighborhood_index
}
\description{
Socioeconomic indicators by neighborhood
}
\keyword{datasets}
